feat(lsp-tool): split into multiple files
